api/requirements.txt
fastapi[standard]>=0.115.12
python-dotenv
pydantic-settings
mlflow==2.22.0
numpy==1.26.4
gensim==4.3.3
tensorflow-text==2.18.1
tensorflow==2.18.0
nltk
tensorflow_hub

scikit-learn==1.6.1

# OpenTelemetry core
opentelemetry-api>=1.34.1
opentelemetry-sdk>=1.34.1
# instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests

# Azure exporters
azure-monitor-opentelemetry-exporter>=1.0.0b37